Reverted to Electron 11.3.0 to fix https://github.com/APTrust/dart/issues/421
Bug https://github.com/APTrust/dart/issues/421, which is caused by context isolation changes in Electron 12.0.0. Neither of the recommended fixes work. The "easy" and unsafe fix of disabling contextIsolation fails with same error we see when contextIsolation is enabled ("require() is not defined"). The proper fix of including a preload script and exposing a limited set of objects and functions through exposeInMainWorld does not work at all. Electron neither exposes the items nor reports errors explaining why the items were not exposed.